+++ model/interface.go
@@ -12,11 +12,11 @@ type VerificationService interface {
 	GetVersion() Version
 
 	// Verify takes names-strings and options and returns verification result.
-	Verify(VerifyParams) []Verification
+	Verify(VerifyParams) []*Verification
 
 	// GetDataSources takes data-source id and opts and returns the data-source
 	// metadata.  If no id is provided, it returns metadata for all data-sources.
-	GetDataSources(DataSourcesOpts) []DataSource
+	GetDataSources(DataSourcesOpts) []*DataSource
 
 	// GetPort returns port of the HTTP/1 service.
 	GetPort() int